Add implied assertion in WheelInfo
Some simple indexes (anaconda), only list the full-pathname of download
URL (assuming same domain, which make sens).

WheelInfo.download does not work on this case at it tries to make the
requests on the current page domain.

Thus we implicitely assume that URL start with http (well https:// would
be better, but when you prototype local index could be http.)

The error can be weird if we let it propagate (bad ziplife as you try to
decode often a 404 html page as ZIP).

This thus just add an assert at construction time to catch the error
early and that we start with actual protocol.

It should in the end be pushed earler in the code (likely at parsing
time), where we are likely to know the index URL and be able to resolve
URLs at that time.

I think there was also a missing comma in test parametrisations, and I
added a !r in a few places as I had some error with empty filenames,
without quotes, so hard to see.

The conftest was also update to explicitely pass a fake URL instead of
just a filename.

I'm not entirely sure we should allow constructing URLs with things that
don't have a protocol though, as this can be confusing.